+++ Ameko/DataModels/ShiftTimesDirection.cs
@@ -1,0 +1,19 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+namespace Ameko.DataModels;
+
+/// <summary>
+/// Direction to shift times in
+/// </summary>
+public enum ShiftTimesDirection
+{
+    /// <summary>
+    /// Shift times forward
+    /// </summary>
+    Forward,
+
+    /// <summary>
+    /// Shift times backward
+    /// </summary>
+    Backward,
+}
