+++ Ameko/DataModels/ShiftTimesType.cs
@@ -1,0 +1,19 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+namespace Ameko.DataModels;
+
+/// <summary>
+/// Type of shifting
+/// </summary>
+public enum ShiftTimesType
+{
+    /// <summary>
+    /// Shift times by timestamp
+    /// </summary>
+    Time,
+
+    /// <summary>
+    /// Shift times by frame count
+    /// </summary>
+    Frames,
+}
