+++ Holo/IO/Paths.cs
@@ -7,16 +7,15 @@ public static class Paths
     /// <summary>
     /// Default file location for the application configuration
     /// </summary>
-    public static Uri Configuration =>
-        new(Path.Combine(DirectoryService.ConfigHome, "config.json"));
+    public static Uri Configuration => new(Path.Combine(Directories.ConfigHome, "config.json"));
 
     /// <summary>
     /// Default file location for globally-accessible objects
     /// </summary>
-    public static Uri Globals => new(Path.Combine(DirectoryService.ConfigHome, "globals.json"));
+    public static Uri Globals => new(Path.Combine(Directories.ConfigHome, "globals.json"));
 
     /// <summary>
     /// Default file location for keybinds
     /// </summary>
-    public static Uri Keybinds => new(Path.Combine(DirectoryService.ConfigHome, "keybinds.json"));
+    public static Uri Keybinds => new(Path.Combine(Directories.ConfigHome, "keybinds.json"));
 }
